+++ bin/mbtiles-server.js
@@ -3,7 +3,7 @@ const meow = require('meow')
 const updateNotifier = require('update-notifier')
 const pkg = require('../package.json')
 const server = require('../index')
-const config = require('../config')
+const DEFAULT = require('../config')
 
 // Update if required
 updateNotifier({pkg}).notify()
@@ -27,11 +27,11 @@ const cli = meow(`
 })
 
 // Define default options
-const cache = cli.flags.cache || config.CACHE
-const domain = cli.flags.domain || config.DOMAIN
-const port = cli.flags.port || config.PORT
-const protocol = cli.flags.protocol || config.PROTOCOL
-const verbose = cli.flags.verbose || config.VERBOSE
+const cache = cli.flags.cache || DEFAULT.CACHE
+const domain = cli.flags.domain || DEFAULT.DOMAIN
+const port = cli.flags.port || DEFAULT.PORT
+const protocol = cli.flags.protocol || DEFAULT.PROTOCOL
+const verbose = cli.flags.verbose || DEFAULT.VERBOSE
 
 // Verbose output
 const status = `